Код разбит на функции + константы вынесены в отдельный модуль
