+++ src/utils/constants.js
@@ -1,0 +1,28 @@
+const TASKS_COUNT = 19;
+const TASK_STEP = 8;
+
+export {TASKS_COUNT, TASK_STEP};
+
+const HASHTAGS = [
+  `homework`,
+  `theory`,
+  `practice`,
+  `intensive`,
+  `keks`
+];
+
+const DESCRIPTIONS = [
+  `Изучить теорию`,
+  `Сделать домашку`,
+  `Пройти интенсив на соточку`
+];
+
+const COLORS = [
+  `black`,
+  `yellow`,
+  `blue`,
+  `green`,
+  `pink`
+];
+
+export {HASHTAGS, DESCRIPTIONS, COLORS};
